Reduce duplicated code in tests
