+++ src/cache/__tests__/FifoMapCacheObject.spec.js
@@ -1,90 +1,20 @@
-import FifoMapCacheObject from '../FifoMapCacheObject';
-import * as validateCacheSize from '../util/validateCacheSize';
-
-function newCache(cacheSize) {
-  return new FifoMapCacheObject({cacheSize});
-}
-
-function fillCache(cache, entries = []) {
-  entries.forEach(entry => cache.set(entry, entry));
-  return cache;
-}
+import CacheObject from '../FifoMapCacheObject';
+import testFifoBehavior from '../__util__/testFifoBehavior';
+import testBasicBehavior from '../__util__/testBasicBehavior';
+import fillCacheWith from '../__util__/fillCacheWith';
 
 describe('FifoMapCacheObject', () => {
-  it('Should return cached value', () => {
-    const cache = newCache(5);
-    const actual = () => {};
-    const cacheKey = {};
+  testBasicBehavior(CacheObject, {cacheSize: 10});
+  testFifoBehavior(CacheObject);
 
-    cache.set(cacheKey, actual);
-    const expected = cache.get(cacheKey);
+  it('Should handle any kind of cache key', () => {
+    const cache = new CacheObject({cacheSize: 5});
+    const entries = new Set([1, {}, 3, [], null]);
 
-    expect(actual).toBe(expected);
-  });
+    fillCacheWith(cache, entries);
 
-  it('Should limit cache queue by removing the first added items', () => {
-    const cache = newCache(5);
-    const entries = [1, 2, 3, 4];
-    const newEntries = [5, 6, 7];
-
-    fillCache(cache, entries);
-    fillCache(cache, newEntries);
-
-    expect(cache.get(1)).toBe(undefined);
-    expect(cache.get(2)).toBe(undefined);
-    [4, 5, 6, 7].forEach(entry => {
+    entries.forEach(entry => {
       expect(cache.get(entry)).toBe(entry);
     });
-  });
-
-  it('Should remove a single item', () => {
-    const cache = newCache(5);
-    const entries = [1, 2, 3, 4, 5];
-    fillCache(cache, entries);
-
-    cache.remove(3);
-
-    expect(cache.get(3)).toBe(undefined);
-    [1, 2, 4, 5].forEach(entry => {
-      expect(cache.get(entry)).toBe(entry);
-    });
-  });
-
-  it('Should mantain cache updated after removing extraneous entry', () => {
-    const cache = newCache(5);
-    const entries = [1, 2, 3, 4, 5];
-    fillCache(cache, entries);
-
-    cache.remove(7); // Extraneous
-    cache.remove(3);
-    cache.set(6, 6);
-    cache.set(7, 7);
-
-    expect(cache.get(1)).toBe(undefined);
-    expect(cache.get(3)).toBe(undefined);
-
-    [2, 4, 5, 6, 7].forEach(entry => {
-      expect(cache.get(entry)).toBe(entry);
-    });
-  });
-
-  it('Should clear the cache', () => {
-    const cache = newCache(5);
-    const entries = [1, 2, 3, 4, 5];
-    fillCache(cache, entries);
-
-    cache.clear();
-
-    [1, 2, 3, 4, 5].forEach(entry => {
-      expect(cache.get(entry)).toBe(undefined);
-    });
-  });
-
-  it('Should validate `cacheSize` parameter', () => {
-    const spy = jest.spyOn(validateCacheSize, 'default');
-    newCache(5);
-
-    expect(spy).toHaveBeenCalledWith(5);
-    spy.mockRestore();
   });
 });
