+++ src/cache/__tests__/FlatMapCacheObject.spec.js
@@ -1,45 +1,18 @@
-import FlatMapCacheObject from '../FlatMapCacheObject';
-
-function fillCache(cache, entries = []) {
-  entries.forEach(entry => cache.set(entry, entry));
-  return cache;
-}
+import CacheObject from '../FlatMapCacheObject';
+import testBasicBehavior from '../__util__/testBasicBehavior';
+import fillCacheWith from '../__util__/fillCacheWith';
 
 describe('FlatMapCacheObject', () => {
-  it('Should return cached value', () => {
-    const cache = new FlatMapCacheObject();
-    const actual = () => {};
-    const cacheKey = {};
+  testBasicBehavior(CacheObject);
 
-    cache.set(cacheKey, actual);
-    const expected = cache.get(cacheKey);
-
-    expect(actual).toBe(expected);
-  });
-
-  it('Should remove a single item', () => {
-    const cache = new FlatMapCacheObject();
+  it('Should handle any kind of cache key', () => {
+    const cache = new CacheObject(CacheObject);
     const entries = new Set([1, {}, 3, [], null]);
-    fillCache(cache, entries);
 
-    entries.delete(3);
-    cache.remove(3);
-
-    expect(cache.get(3)).toBe(undefined);
-    entries.forEach(entry => {
-      expect(cache.get(entry)).toEqual(entry);
-    });
-  });
-
-  it('Should clear the cache', () => {
-    const cache = new FlatMapCacheObject();
-    const entries = new Set([1, {}, 3, [], null]);
-    fillCache(cache, entries);
-
-    cache.clear();
+    fillCacheWith(cache, entries);
 
     entries.forEach(entry => {
-      expect(cache.get(entry)).toBe(undefined);
+      expect(cache.get(entry)).toBe(entry);
     });
   });
 });
