+++ src/cache/__util__/fillCacheWith.js
@@ -1,0 +1,6 @@
+function fillCacheWith(cache, entries = []) {
+  entries.forEach(entry => cache.set(entry, entry));
+  return cache;
+}
+
+export default fillCacheWith;
